Ishaan/creating service

+++ order-inventory/src/main/java/com/spring/order_inventory/repository/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
-    List<Order> findByCustomerCustomerId(Long id);
-    List<Order> findByStoreStoreId(Long id);
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+    List<Order> findByCustomerCustomerId(Integer id);
+    List<Order> findByStoreStoreId(Integer id);
 }
