+++ order-inventory/src/main/java/com/spring/order_inventory/repository/StoreRepository.java
@@ -3,4 +3,4 @@ package com.spring.order_inventory.repository;
 import com.spring.order_inventory.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Long> {}
+public interface StoreRepository extends JpaRepository<Store, Integer> {}
